# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "<a href='http://cse341-project2-suggs.herokuapp.com/auth/logout'>Logout</a>",
    "version": "1.0.0"
  },
  "host": "cse341-project2-suggs.herokuapp.com",
  "basePath": "/",
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/api-docs": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/bucketlist/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "item name"
                },
                "deadline": {
                  "example": "\"2999\""
                },
                "priority": {
                  "example": "\"99\""
                },
                "description": {
                  "example": "a description of the goal"
                },
                "plan": {
                  "example": "a plan to obtain the goal"
                },
                "links": {
                  "example": "reference links (if any)"
                },
                "notes": {
                  "example": "comments or notes (if any)"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/bucketlist/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "item name"
                },
                "deadline": {
                  "example": "\"2999\""
                },
                "priority": {
                  "example": "\"99\""
                },
                "description": {
                  "example": "a description of the goal"
                },
                "plan": {
                  "example": "a plan to obtain the goal"
                },
                "links": {
                  "example": "reference links (if any)"
                },
                "notes": {
                  "example": "comments or notes (if any)"
                }
              }
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}